dags/requirements.txt
requests
selenium
selenium-page-factory
psycopg2-binary
pydantic
apache-airflow-providers-postgres
minio